+++ Sources/Classes/DocumentVerification/Model/AutoCapture.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public enum AutoCapture {
+  case autoCapture
+  case autoCaptureOnly
+  case manualCaptureOnly
+}