+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 * Removed `AntiFraud` response in `JobStatus` calls
 * Removed the default `ConsentInformation` set when no consent is passed
+* Removed the default `ConsentInformation`
+* Changed `enableAutoCapture` to `AutoCapture` enum to allow to allow partners change document capture options
 
 ## 11.0.2 - July 10, 2025